Jeremy-
Sorry for the delay. We are about 75-80% done. Please find attached the
zip
file containing the tex file and figures. Some of the tex stuff may lead
to
warnings if compiled using pdflatex. We will fix these soon.

If you can upload the tex file in the inria page it will be great.

-tx
dibyendu

(See attached file: For Jeremy.zip)



git-svn-id: svn://scm.gforge.inria.fr/svnroot/ssabook@84 9a4b08f2-1d31-4dde-917c-3c42c574bc2c